Add empty spans to prevent auto bug-linking.
Github interprets the # symbol specially as a reference to an issue,
but Sourceforge comments are almost certainly not referring to
Github issues. This can cause problems when comments contain eg.
gdb stack traces, which use the same syntax for frame numbers.
Detect when comments do this and add empty spans to prevent linking.
This fixes #2.